requirements.txt
requests
django
djangorestframework
fabric
pymongo
